refactor(schema): modularize style schema
Split the style schema monolith into private modules.

Preserve public re-exports and serde wire format.
